requirements.txt
compoundfiles
compressed-rtf
extract-msg==0.48.0
mail-parser
bs4
lxml
eml-parser
multidecoder>=1.0.0,<2
chardet
dnspython==2.6.1
